refactor(harness): extract settings-manager into harness
Move settings load/save implementation from src/config.ts to src/harness/settings-manager.ts. src/config.ts becomes a thin re-export layer for backward compatibility. No behavior change.
